CRMS-754 Add non-empty check to avoide notice error

+++ application/libraries/Invoice_lib.php
@@ -217,7 +217,7 @@ class Invoice_lib {
         if(file_exists($sign_path)){
             unlink($sign_path);
         }
-        if(file_exists($stamp_path)){
+        if(!empty($stamp_path) && file_exists($stamp_path)){
             unlink($stamp_path);
         }
